+++ src/controllers/commands/CommandController.cpp
@@ -391,6 +391,8 @@ CommandController::CommandController(const Paths &paths)
     this->registerCommand("/clearmessages", &commands::clearmessages);
     this->registerCommand("/clearsystemmessages",
                           &commands::clearSystemMessages);
+    this->registerCommand("/clearallsystemmessages",
+                          &commands::clearAllSystemMessages);
 
     this->registerCommand("/settitle", &commands::setTitle);