+++ src/controllers/commands/builtin/Misc.cpp
@@ -10,6 +10,7 @@
 #include "controllers/commands/CommandContext.hpp"
 #include "controllers/userdata/UserDataController.hpp"
 #include "providers/kick/KickChannel.hpp"
+#include "providers/kick/KickChatServer.hpp"
 #include "providers/twitch/api/Helix.hpp"
 #include "providers/twitch/TwitchAccount.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
@@ -469,6 +470,16 @@ QString clearSystemMessages(const CommandContext &ctx)
         return "";
     }
     ctx.channel->clearSystemMessages();
+    return "";
+}
+
+QString clearAllSystemMessages(const CommandContext &ctx)
+{
+    (void)ctx;
+    getApp()->getTwitch()->forEachChannelAndSpecialChannels(
+        [](const ChannelPtr &chan) { chan->clearSystemMessages(); });
+    getApp()->getKickChatServer()->forEachChannel(
+        [](KickChannel &chan) { chan.clearSystemMessages(); });
     return "";
 }