+++ src/controllers/commands/builtin/Misc.hpp
@@ -27,6 +27,7 @@ QString popout(const CommandContext &ctx);
 QString popup(const CommandContext &ctx);
 QString clearmessages(const CommandContext &ctx);
 QString clearSystemMessages(const CommandContext &ctx);
+QString clearAllSystemMessages(const CommandContext &ctx);
 QString openURL(const CommandContext &ctx);
 QString sendRawMessage(const CommandContext &ctx);
 QString injectFakeMessage(const CommandContext &ctx);